python/requirements.txt
argparse==1.2.1
distribute==0.6.24
gevent==0.13.8
greenlet==0.4.0
websocket-client==0.9.0
wsgiref==0.1.2
